requirements.txt
asgiref==3.5.2
Babel==2.10.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.6.15
cloudinary==1.29.0
dj-database-url==0.5.0
Django==4.0.5
django-bootstrap-v5==1.0.11
django-bootstrap5==21.3
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-money==3.0.0
django-tinymce==3.4.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
py-moneyed==2.0
python-decouple==3.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
typing-extensions==4.3.0
urllib3==1.26.9
whitenoise==6.2.0
